adding svag support - fixing implicit warning on vs2005 for ps2_xxx functions
git-svn-id: https://vgmstream.svn.sourceforge.net/svnroot/vgmstream@114 51a99a44-fe44-0410-b1ba-c3e57ba2b86b

+++ src/meta/ps2_ads.c
@@ -79,7 +79,7 @@ VGMSTREAM * init_vgmstream_ps2_ads(const char * const filename) {
 
             vgmstream->ch[i].channel_start_offset=
                 vgmstream->ch[i].offset=
-                0x28+vgmstream->interleave_block_size*i;
+                (off_t)(0x28+vgmstream->interleave_block_size*i);
         }
     }
